Ready For you to edit.
Once you have it on your computer, delete this file from GitHub so I will know you are currently modifying it. When you are done editing for the day, simply upload the filed back here and say "Ready For you to edit" and then I will download it and edit it. If for some reason, I am editing it, and you realize you need to make a change or add something, just add it to the backup copy on your desktop and then when its your turn simply add the changes. This should reduce some of the clashes we may run into while we try to work on this diagram from home.