click me
I have organized the classes and I had to make a lot of changes to the associations we were using after yesterdays lecture. There is a lot more to add so I will revist it when you are done editing. Just shoot me an email when you are done editing and I will pick up where you left off. I also left a lot of comments on classes.